requirements.txt
# Combined requirements for mono-repo project
# Consolidated from root requirements.txt and build/requirements/*.txt

# ============================================================================
# CORE DEPENDENCIES (Base)
# ============================================================================

# Configuration and Environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
psycopg2==2.9.9  # Production optimized version
snowflake-sqlalchemy==1.5.1
snowflake-connector-python==3.6.0
snowflake-connector-python[pandas]==3.6.0  # UAT enhanced version
alembic==1.12.1

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Utilities
requests==2.31.0
click==8.1.7
python-dateutil>=2.8.0

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# ============================================================================
# WEB APPLICATIONS
# ============================================================================

# Dash Applications
dash==2.14.2
plotly==5.17.0

# Legacy Web Framework (if needed)
flask>=3.0.0

# ============================================================================
# DATA PIPELINE & ORCHESTRATION
# ============================================================================

# Airflow
apache-airflow==2.7.3

# ============================================================================
# TESTING & QUALITY ASSURANCE
# ============================================================================

# Core Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Load Testing (UAT)
locust==2.17.0

# Testing Utilities (Dev)
factory-boy==3.3.0
faker==20.1.0
httpx==0.25.2

# ============================================================================
# DEVELOPMENT TOOLS
# ============================================================================

# Jupyter Environment
jupyter==1.0.0
ipython==8.17.2
notebook==7.0.6

# Debugging
ipdb==0.13.13
pdb++==0.10.3

# Development Server
watchdog==3.0.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# ============================================================================
# PRODUCTION & DEPLOYMENT
# ============================================================================

# Production WSGI Server
gunicorn==21.2.0

# Kubernetes
kubernetes>=28.1.0

# ============================================================================
# SECURITY & AUTHENTICATION
# ============================================================================

# Security
cryptography==41.0.8
pyjwt>=2.8.0

# ============================================================================
# CONFIGURATION & VALIDATION
# ============================================================================

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ============================================================================
# CLOUD SERVICES (Conditional - Uncomment as needed)
# ============================================================================

# AWS SDK (uncomment for environments with AWS access)
# boto3==1.34.0
# botocore==1.34.0

# ============================================================================
# ENVIRONMENT-SPECIFIC NOTES
# ============================================================================

# This file combines requirements from:
# - Root requirements.txt (general dependencies)
# - build/requirements/base.txt (core application dependencies)
# - build/requirements/dev.txt (development tools)
# - build/requirements/uat.txt (UAT testing tools)
# - build/requirements/prod.txt (production optimizations)
#
# For environment-specific installations:
# - Development: Install all packages
# - UAT: Install base + monitoring + testing tools
# - Production: Install base + production optimizations only
